vita: get the 2D UI rendering
The port reached its main loop but drew nothing except the frame clear
colour, with no GL error anywhere. Two independent silent failures, both
in irrlicht's GLES2 driver.

1. Every 2D draw call was discarded.

COGLES2Driver's useCoreContext defaults to true and is only ever assigned
in the EGL constructor, from EglContext->isLegacyDevice(). The SDL device
used on Vita never touches it, so it kept the default and each
useCoreContext guard took its early return - including the one at the top
of setRenderStates2DMode(), which is where the whole GUI enters the
driver. glClear is unaffected by any of this, hence a correct full screen
clear with zero geometry and not one GL error to point at it.

vitaGL reports GLSL ES 1.00, so STK selects its fixed pipeline
(CVS->isGLSL() == false) and deliberately routes 2D through irrlicht,
which then did nothing with it.

Re-enable only the three state setters the 2D path calls:
setBasicRenderStates (from OnSetMaterial), setTextureRenderStates (from
OnRender) and setRenderStates2DMode itself. All three use core GLES2 only
- glEnable/Disable, glBlendEquation, glColorMask, glTexParameteri.
setRenderStates3DMode is left bailing on purpose; enabling every guard at
once drove vitaGL's fixed-function emulation into a SceGxm data abort, so
3D stays unfixed here and gets its own change.

2. Nothing was textured.

GXP has no bool type, so vitaGL translates a GLSL bool uniform and
reports it as GL_FLOAT - GL_BOOL appears nowhere in vitaGL. Irrlicht
uploads uUseTexture through setPixelShaderConstant(s32*), whose switch
handles GL_INT/GL_BOOL and samplers but has no GL_FLOAT case, so it fell
to "default: status = false" and never uploaded anything. The return
value is ignored at the call site. uUseTexture therefore stayed 0,
texture2D was never sampled, and every surface drew flat vertex colour
with text as solid squares.

Declare uUseTexture as float and upload it as one. The same fix is
applied pre-emptively to uTextureUsage0 in COGLES2FixedPipeline.fsh,
which is the identical bug and would produce untextured 3D. This is not
Vita-specific - the shaders are shared data files - but it is
behaviourally identical on every backend: irrlicht's float path handles
GL_FLOAT everywhere, and comparing a float against 0.5 is valid GLSL ES.

3. Music streaming overflowed its thread stack.

MusicOggStream::streamIntoBuffer() opened with a variable length array of
m_buffer_size bytes - 44100 for vorbis, 192000 with Opus - on the SFX
manager thread, which std::thread gives a default stack. It faulted on a
store to its own stack before doing any work. Move the scratch buffer
into the object; the function prologue drops from 44100 bytes of stack to
36. This is a latent bug on any platform with small thread stacks, not
just Vita.

Also record, in irr_driver.cpp, why vglSetSemanticBindingMode() must not
be called: VGL_MODE_POSTPONED is the only mode that honours
glBindAttribLocation() between compile and link, which is exactly where
COGLES2MaterialRenderer::init binds. Switching away from it silently
discards every attribute binding.

2D UI, text and textures now render. 3D still does not.

+++ lib/irrlicht/source/Irrlicht/COGLES2Renderer2D.cpp
@@ -80,7 +80,10 @@ bool COGLES2Renderer2D::OnRender(IMaterialRendererServices* service, E_VERTEX_TY
 		RenderTargetSize = renderTargetSize;
 	}
 
-	int UseTexture = Texture ? 1 : 0;
+	// Uploaded as a float: the shader declares uUseTexture as float because
+	// GXP has no bool type (vitaGL reports GL_FLOAT), and irrlicht's integer
+	// upload path has no GL_FLOAT case so it would silently drop the value.
+	float UseTexture = Texture ? 1.0f : 0.0f;
 	setPixelShaderConstant(UseTextureID, &UseTexture, 1);
 
 	return true;
